add abstract class to Genre

+++ MusicGradeApp/MusicGradeApp/GenreBase.cs
@@ -17,7 +17,7 @@ namespace MusicGradeApp
             MusicGenre = musicGenre;
         }
 
-        public abstract void AddTrack(string title, float rating);
+        public abstract void AddTrack(string title, int rating);
         public abstract void AddTrack(string title, string rating);
         public abstract void ShowTracks();
         public abstract Statistics GetStatistics();
